ensemble better, f1 score
so the thing is that i actually only have two ground truth malicious logins. we need to split by 50% so one is in each for now at least. but that's a really interesting consideration. the neural network can't train until there actually is a malicious login, ideally many. i may have to actually come up with a way to add synthetic data that's high risk, probably using statistics once again